operations/sagemaker-mlflow-trace-evaluate-langgraph-agent/pyproject.toml
[project]
name = "agentops-traceevaluate-langgraph-mlflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dask>=2025.2.0",
    "evaluate>=0.4.3",
    "ipykernel>=6.29.5",
    "langchain-aws>=0.2.14",
    "langgraph>=0.3.2",
    "mlflow>=2.20.2",
    "nltk>=3.9.1",
    "psutil>=7.0.0",
    "python-dotenv>=1.0.1",
    "ragas==0.2.13",
    "rouge-score>=0.1.2",
    "sagemaker-mlflow>=0.1.0",
    "textstat>=0.7.5",
    "tf-keras>=2.15.0",
    "tiktoken>=0.9.0",
    "torch>=2.6.0",
    "tqdm>=4.67.1",
    "transformers>=4.49.0",
]
